requirements/base.txt
celery
dj-database-url
django
django-allauth
django-ckeditor
django-cleanup
django-cmstemplates
django-codemirror-widget
django-el-pagination
django-widget-tweaks
django-embed-video
easy-thumbnails
envvars
hiredis
ipython
lxml  # libxml2-dev libxslt-dev
pillow  # libjpeg8 libjpeg8-dev libfreetype6 libfreetype6-dev zlib1g-dev
psycopg2  # libpq-dev
pymorphy2-dicts-ru
pymorphy2[fast]
pytils
redis
rules
